style: implement CourseFilters component for enhanced course filtering options
